pyproject.toml
[tool.poetry]
package-mode = false
name = "lappis"
version = "0.1.0"
description = "Gestão de Dados do IPEA - Aplicações Python"
authors = ["Lappis UNB"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.11"
apache-airflow = "2.8.1"
apache-airflow-providers-postgres = "5.14.0"
dbt-core = "*"
dbt-postgres = "*"
pandas = "*"
requests = "*"
sqlalchemy = "*"
zeep = "*"
imap-tools = "*"

[tool.poetry.group.dev.dependencies]
black = "*"
ruff = "*"
pytest = "*"
pytest-cov = "*"
mypy = "*"
types-psycopg2 = "*"
pandas-stubs = "*"
jupyter = "*"
shandy-sqlfmt = "*"
sqlfluff = "*"
sqlfluff-templater-dbt = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 90
target-version = ["py311"]
skip-string-normalization = false
include = "\\.pyi?$"
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 90
target-version = "py311"
exclude = [
    ".git",
    ".venv",
    "migrations",
    "build",
    "dist",
    "*.pyi",
    ".ipynb_checkpoints",
    "*.ipynb"
]

[tool.ruff.lint]
select = ["E", "F", "W", "C90"]
ignore = []

[tool.mypy]
python_version = "3.11"
files = ["**/*.py"]
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
disable_error_code = ["arg-type"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.sqlfmt]
line_length = 90
dialect = "polyglot"
exclude = [
    ".git/",
    ".venv/",
    "build/",
    "dist/",
]

[tool.sqlfluff.core]
dialect = "postgres"
templater = "dbt"
max_line_length = 90
exclude_rules = "L016"

[tool.sqlfluff.indentation]
indented_joins = true
indented_using_on = true
template_blocks_indent = true

[tool.sqlfluff.rules.capitalisation.keywords]
capitalisation_policy = "lower"

[tool.sqlfluff.rules.capitalisation.identifiers]
capitalisation_policy = "lower"

[tool.sqlfluff.rules.capitalisation.functions]
extended_capitalisation_policy = "lower"

[tool.sqlfluff.rules.capitalisation.literals]
capitalisation_policy = "lower"

[tool.sqlfluff.templater]
unwrap_wrapped_queries = true

[tool.sqlfluff.templater.dbt]
project_dir = "."
profiles_dir = "."
profile = "default"
target = "dev"
